+++ implementation/js/src/infrastructure/lib+-/backbone/backbone-deep-model.js
@@ -1721,22 +1721,23 @@ var merge = require('lodash.merge');
  * @param  {Object}      Nested object e.g. { level1: { level2: 'value' } }
  * @return {Object}      Shallow object with path names e.g. { 'level1.level2': 'value' }
  */
-function objToPaths(obj, ignoreArray) { //Tim's Hack: added ignoreArray option!
+function objToPaths(obj, ignoreArray, lvlCap) { //Tim's Hack: added ignoreArray, lvlCap options!
 	var ret = {},
 		separator = DeepModel.keyPathSeparator;
 
 	for (var key in obj) {
 		var val = obj[key];
 
-		if (val && (val.constructor === Object || (!ignoreArray && val.constructor === Array)) && !_.isEmpty(val)) {
-			//Recursion for embedded objects
-			var obj2 = objToPaths(val, ignoreArray);
+		if ((lvlCap === undefined || lvlCap > 0) && val && (val.constructor === Object || (!ignoreArray && val.constructor === Array)) && !_.isEmpty(val)) {
+        //Recursion for embedded objects
+  			var obj2 = objToPaths(val, ignoreArray, _.isNumber(lvlCap)? lvlCap - 1 : lvlCap);
 
-			for (var key2 in obj2) {
-				var val2 = obj2[key2];
+  			for (var key2 in obj2) {
+  				var val2 = obj2[key2];
 
-				ret[key + separator + key2] = val2;
-			}
+  				ret[key + separator + key2] = val2;
+  			}
+
 		} else {
 			ret[key] = val;
 		}
@@ -1860,7 +1861,7 @@ var DeepModel = Backbone.Model.extend({
 	// Override set
 	// Supports nested attributes via the syntax 'obj.attr' e.g. 'author.user.name'
 	set: function(key, val, options) {
-		var attr, attrs, unset, changes, silent, changing, prev, current;
+		var attr, attrs, unset, changes, silent, replace, changing, prev, current;
 		if (key == null) return this;
 
 		// Handle both `"key", value` and `{key: value}` -style arguments.
@@ -1879,6 +1880,7 @@ var DeepModel = Backbone.Model.extend({
 		// Extract attributes and options.
 		unset = options.unset;
 		silent = options.silent;
+    replace = options.replace; //true, or the level of depth that replacement need to happen.
 		changes = [];
 		changing = this._changing;
 		this._changing = true;
@@ -1893,10 +1895,15 @@ var DeepModel = Backbone.Model.extend({
 		if (this.idAttribute in attrs) this.id = attrs[this.idAttribute];
 
 		//<custom code>
-		attrs = objToPaths(attrs, true);//Tim's Hack: activate ignoreArray option! no more array.0.xyz
-                                    //This is to fix the array 'shrink' problem with 'change':
-                                    //set('array', [1, 2, 3])
-                                    //set('array', [1]) will give no 'change' event.
+		attrs = objToPaths(attrs, true, replace? (_.isNumber(replace)? replace: 0) : undefined);
+    //Tim's Hack: 
+    //  1. Always use ignoreArray option! no more array.0.xyz
+    //     This is to fix the array 'shrink' not emitting 'change' problem:
+    //     set('array', [1, 2, 3])
+    //     set('array', [1]) would give no 'change' event previousely.
+    //     
+    //  2. Always honor the replace option! it can be `true` (indicates lvl 0) or a 1 ~ n level number
+    //     This is to allow finer control over the value replace/merge depth when given val is an object.
 		//</custom code>
 
 		// For each `set` attribute, update or delete the current value.